st190x: break out MIML parsing to separate classes
