+++ miml-maven-plugin/src/test/java/org/jmisb/maven/ClassEntryModelTest.java
@@ -11,44 +11,38 @@ public class ClassEntryModelTest {
 
     @Test
     public void minValue() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         assertEquals(entry.getMinValue(), 0.0);
         assertNull(entry.getMaxValue());
     }
 
     @Test
     public void number() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         assertEquals(entry.getNumber(), 38);
     }
 
     @Test
     public void name() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         assertEquals(entry.getName(), "syncPulseFreq");
     }
 
     @Test
     public void nameSentenceCase() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         assertEquals(entry.getNameSentenceCase(), "SyncPulseFreq");
     }
 
     @Test
     public void type() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         assertEquals(entry.getTypeName(), "Real");
     }
 
     @Test
     public void units() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         assertEquals(entry.getUnits(), "Hz");
         assertEquals(entry.getEscapedUnits(), "Hz");
     }
@@ -56,8 +50,7 @@ public class ClassEntryModelTest {
     @Test
     public void escapedUnits() {
         ClassModelEntry entry =
-                MimlToJava.parseClassEntry(
-                        "34_relativeHumidity     : Real (0, 100, 0.1)      {  %}; ");
+                Parser.parseClassEntry("34_relativeHumidity     : Real (0, 100, 0.1)      {  %}; ");
         assertEquals(entry.getUnits(), "%");
         assertEquals(entry.getEscapedUnits(), "%%");
     }
@@ -65,8 +58,7 @@ public class ClassEntryModelTest {
     @Test
     public void realTraits() {
         ClassModelEntry entry =
-                MimlToJava.parseClassEntry(
-                        "34_relativeHumidity     : Real (0, 100, 0.1)      {  %}; ");
+                Parser.parseClassEntry("34_relativeHumidity     : Real (0, 100, 0.1)      {  %}; ");
         assertEquals(entry.getMinValue(), 0.0, 0.0000000000001);
         assertEquals(entry.getMaxValue(), 100.0, 0.0000000000001);
         assertEquals(entry.getResolution(), 0.1, 0.0000000000001);
@@ -75,7 +67,7 @@ public class ClassEntryModelTest {
     @Test
     public void maxValue() {
         ClassModelEntry entry =
-                MimlToJava.parseClassEntry(
+                Parser.parseClassEntry(
                         "  36_differentialPressure : Real (0, 5000)          {hPa};");
         assertEquals(entry.getMinValue(), 0.0);
         assertEquals(entry.getMaxValue(), 5000.0);
@@ -84,7 +76,7 @@ public class ClassEntryModelTest {
     @Test
     public void maxValue2PI() {
         ClassModelEntry entry =
-                MimlToJava.parseClassEntry(
+                Parser.parseClassEntry(
                         "  38_windDirection        : Real (0, TWO_PI)        {rad};");
         assertEquals(entry.getMinValue(), 0.0);
         assertEquals(entry.getMaxValue(), 6.283185307179586, 0.000000000000001);
@@ -93,8 +85,7 @@ public class ClassEntryModelTest {
     @Test
     public void maxValuePI() {
         ClassModelEntry entry =
-                MimlToJava.parseClassEntry(
-                        "  38_windDirection        : Real (0, PI)        {rad};");
+                Parser.parseClassEntry("  38_windDirection        : Real (0, PI)        {rad};");
         assertEquals(entry.getMinValue(), 0.0);
         assertEquals(entry.getMaxValue(), 3.141592653589793, 0.000000000000001);
     }
@@ -102,7 +93,7 @@ public class ClassEntryModelTest {
     @Test
     public void maxValueHALF_PI() {
         ClassModelEntry entry =
-                MimlToJava.parseClassEntry(
+                Parser.parseClassEntry(
                         "  38_windDirection        : Real (0, HALF_PI)        {rad};");
         assertEquals(entry.getMinValue(), 0.0);
         assertEquals(entry.getMaxValue(), 1.5707963267948965, 0.000000000000001);
@@ -110,7 +101,7 @@ public class ClassEntryModelTest {
 
     @Test
     public void REF() {
-        ClassModelEntry entry = MimlToJava.parseClassEntry("  02_timer : REF<Timer> {None};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  02_timer : REF<Timer> {None};  ");
         assertEquals(entry.getNumber(), 2);
         assertEquals(entry.getRefItemType(), "Timer");
         assertEquals(entry.getName(), "timer");
@@ -120,8 +111,7 @@ public class ClassEntryModelTest {
 
     @Test
     public void LIST() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("38_timers : LIST<Timer> (1, *) {None}; ");
+        ClassModelEntry entry = Parser.parseClassEntry("38_timers : LIST<Timer> (1, *) {None}; ");
         assertEquals(entry.getNumber(), 38);
         assertEquals(entry.getListItemType(), "Timer");
         assertEquals(entry.getName(), "timers");
@@ -133,8 +123,7 @@ public class ClassEntryModelTest {
 
     @Test
     public void LIST28() {
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("38_timers : LIST<Timer> (2, 8) {None}; ");
+        ClassModelEntry entry = Parser.parseClassEntry("38_timers : LIST<Timer> (2, 8) {None}; ");
         assertEquals(entry.getNumber(), 38);
         assertEquals(entry.getListItemType(), "Timer");
         assertEquals(entry.getName(), "timers");
@@ -149,8 +138,7 @@ public class ClassEntryModelTest {
         ClassModel parent = new ClassModel();
         parent.setDocument("ST190x");
         parent.setPackageNameBase("test.miml");
-        ClassModelEntry entry =
-                MimlToJava.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
+        ClassModelEntry entry = Parser.parseClassEntry("  38_syncPulseFreq : Real (0.0) { Hz};  ");
         entry.setParent(parent);
         assertEquals(entry.getDocument(), "ST190x");
         assertEquals(entry.getPackageName(), "test.miml.st190x");
